Fix bad design in FeatureWriterStrategy class hierarchy
FileFeatureWriter extends FeatureWriterStrategy.
Both KMLFeatureWriter and ShpFeatureWriter extend FileFeatureWriter.
Yet, FileFeatureWriter implements the writing of shapefiles.

This patch moves that code to ShpFeatureWriter where it belongs.
